lab4/requirements.txt
flask==2.0.1
requests==2.26.0
pycryptodome==3.10.1
